refactor(notification-settings-screen): rename props to types
