+++ app/screens/Menu/NotificationSettings/NotificationSettings.types.ts
