+++ app/screens/Menu/components/MenuSwitchRow/MenuSwitchRow.types.ts
